backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
google-generativeai
python-multipart
python-dotenv
